% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{project}
\alias{project}
\title{Create a projection for matching}
\usage{
project(mod, docs, interactions = FALSE, int.type = c("refit", "phi",
  "theta"))
}
\arguments{
\item{mod}{test}

\item{docs}{test}

\item{interactions}{test}

\item{int.type}{test}
}
\value{
list

\item{mu}{example}
}
\description{
Text about that.
}
\examples{
\dontrun{
example
}
}
\references{
Roberts, M., Stewart, B., Tingley, D., and Airoldi, E. (2013)
"The structural topic model and applied social science." In Advances in
Neural Information Processing Systems Workshop on Topic Models: Computation,
Application, and Evaluation. http://goo.gl/uHkXAQ
}
